Moved packages, drew thicker line between GLObjects and Utilities
