+++ src/main/java/jgloom/common/SharedLibraryLoader.java
@@ -1,4 +1,4 @@
-package jgloom;
+package jgloom.common;
 
 import java.io.File;
 import java.io.FileInputStream;
