+++ src/main/java/jgloom/common/glfw/GLFWWindowContainer.java
@@ -1,5 +1,6 @@
-package jgloom.glfw;
+package jgloom.common.glfw;
 
+import jgloom.glfw.GLFWWindow;
 import org.lwjgl.BufferUtils;
 import org.lwjgl.glfw.GLFW;
 import org.lwjgl.opengl.GL11;
